requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0              # production WSGI server (Railway / VPS)
python-multipart>=0.0.9      # file upload support (template upload)
python-dotenv>=1.0.0

# ── LangGraph / LangChain ─────────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0

# ── Search ────────────────────────────────────────────────────────────────────
exa-py>=1.0.0

# ── Database ──────────────────────────────────────────────────────────────────
motor>=3.3.0                 # async MongoDB driver
pymongo>=4.6.0

# ── Output generation ─────────────────────────────────────────────────────────
reportlab>=4.0.0             # PDF generation

# ── Data / validation ─────────────────────────────────────────────────────────
pydantic>=2.0.0
